ComposeArea: Move extractText to helper function
