+++ src/helpers.ts
@@ -385,3 +385,50 @@ export function copyShallow<T extends object>(obj: T): T {
 export function copyDeep<T extends object>(obj: T): T {
     return JSON.parse(JSON.stringify(obj));
 }
+
+/**
+ * Process a DOM node recursively and extract text from compose area.
+ */
+export function extractText(targetNode: HTMLElement, logWarning: (msg: string) => void, trim = true) {
+    let text = '';
+    const visitChildNodes = (parentNode: HTMLElement) => {
+        // When pressing shift-enter and typing more text:
+        //
+        // - Firefox and chrome insert a <br> between two text nodes
+        // - Safari creates two <div>s without any line break in between
+        //   (except for the first line, which stays plain text)
+        //
+        // Thus, for Safari, we need to detect <div>s and insert a newline.
+
+        // tslint:disable-next-line: prefer-for-of (see #98)
+        for (let i = 0; i < parentNode.childNodes.length; i++) {
+            const node = parentNode.childNodes[i] as HTMLElement;
+            switch (node.nodeType) {
+                case Node.TEXT_NODE:
+                    // Append text, but strip leading and trailing newlines
+                    text += node.nodeValue.replace(/(^[\r\n]*|[\r\n]*$)/g, '');
+                    break;
+                case Node.ELEMENT_NODE:
+                    const tag = node.tagName.toLowerCase();
+                    if (tag === 'div') {
+                        text += '\n';
+                        visitChildNodes(node);
+                        break;
+                    } else if (tag === 'img') {
+                        text += (node as HTMLImageElement).alt;
+                        break;
+                    } else if (tag === 'br') {
+                        text += '\n';
+                        break;
+                    } else if (tag === 'span' && node.hasAttribute('text')) {
+                        text += node.getAttributeNode('text').value;
+                        break;
+                    }
+                default:
+                    logWarning(`Unhandled node: ${node}`);
+            }
+        }
+    };
+    visitChildNodes(targetNode);
+    return trim ? text.trim() : text;
+}
